reafat: reorganizando sistema de pastas
