+++ player/actions.lua
@@ -1,0 +1,3 @@
+local actions = {}
+
+return actions
